### 2018-03-10 #611 @anriseth
Of the existing reviewers, may I suggest @vchuravy and  @ahwillia? They both know Julia very well, and machine learning backgrounds may be helpful in terms of optimization.
Another person in the Julia optimization community that has a an appropriate background is @abelsiqueira.
