### 2018-03-10 #611 @anriseth
The article proof looks fine. I'll also cc @pkofod
