### 2018-03-10 #611 @ahwillia
I'm happy to review.

I've used this package extensively in the past and have had mostly good experiences. I haven't used it in the past year or so.
