### 2018-03-12 #615 @whedon
[REVIEW]: Optim: An optimization package for Julia
**Submitting author:** @anriseth (<a href="http://orcid.org/0000-0002-5861-7885">Asbjørn Nilsen Riseth</a>)
**Repository:** <a href="https://github.com/JuliaNLSolvers/Optim.jl" target ="_blank">https://github.com/JuliaNLSolvers/Optim.jl</a>
**Version:** v0.14.0
