### 2018-03-12 #611 @labarba
Thank you for agreeing to review, @ahwillia — Please go over to the REVIEW issue now, where you can use comments to communicate with the author and editors as you complete your review.
https://github.com/openjournals/joss-reviews/issues/615

@anriseth — In your paper, I do miss some brief note about existing software providing similar functionality (in the same or other programming language). We don't have a novelty requirement, but we do want to see other software mentioned or cited, as appropriate.
