### 2018-03-12 #615 @labarba
@anriseth — In your paper, I do miss some brief note about existing software providing similar functionality (in the same or other programming language). We don't have a novelty requirement, but we do want to see other software mentioned or cited, as appropriate.
