### 2018-03-12 #615 @labarba
Mention other software that has same/similar functionality, briefly comment on differences and equivalence, but no need to discuss or make comparisons requiring you to run the other software. It's a matter of giving acknowledgement to the existence of the other software, not claim originality or superiority.
