### 2018-03-13 #615 @pkofod
> Shouldn’t John Myles White be an author? I’m not familiar with the package history, nor with ownership “rules”, but since he’s made major contributions, it warrants the question. Answering this will probably clarify Tim Holy's situation as well.

Maybe we should bring in @labarba to clarify the rules. The situation is that John Myles White's current employment has very strict rules regarding review of work and publication done outside of his place of work. What we could do is finish these reviews, and then when they "good to go" we can try to see if what is there in the end fits their legal requirements. For good measure, I think I'll ping @johnmyleswhite to show that we're not trying to cheat anyone here.

Regarding @timholy the story is relevant as well, as he initially made a lot of contributions to the package, and is a relevant author as well. It would be :+1: to add him as well, but we were not comfortable adding an author that we have not been able to get explicit consent from.

The two potential co-authors were the two founding persons of Optim, but neither have been active contributors or maintainers for quite a while, and both are aware of this submission. However, @anriseth and I are perfectly happy to include one or both as authors as long as the legality and explicit consent can be established.
