### 2018-03-13 963679a
Add DOIs, clarify mathematical optimization and aim
