### 2018-03-14 0f2d94c
Fixed dep warnings for v0.7.
