+++ src/Optim.jl
@@ -1,6 +1,9 @@
 __precompile__(true)
 
 module Optim
+    using Compat
+    using Compat.LinearAlgebra
+    using Compat.Printf
     using PositiveFactorizations
     using Compat
     using LineSearches
