### 2018-03-14 1027449
JOSS update  (#546)
* Add DOIs, clarify mathematical optimization and aim

* More specific description
