+++ joss/paper.bib
@@ -6,7 +6,9 @@
   number={1},
   pages={65--98},
   year={2017},
-  publisher={SIAM}
+  publisher={SIAM},
+  doi={10.1137/141000671},
+  url={https://doi.org/10.1137/141000671}
 }
 
 @book{mykel2018optimization,
@@ -90,7 +92,7 @@ archivePrefix = "arXiv",
      year = 2018,
     month = jan,
    adsurl = {http://adsabs.harvard.edu/abs/2018arXiv180108572D},
-  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+  adsnote = {Provided by the SAO/NASA Astrophysics Data System},  
 }
 
 @article{dony2018parametric,
@@ -111,7 +113,9 @@ archivePrefix = "arXiv",
   volume={5},
   number={1},
   year={2017},
-  publisher={Ubiquity Press}
+  publisher={Ubiquity Press},
+  url={http://doi.org/10.5334/jors.151},
+  doi={10.5334/jors.151}
 }
 
 @article{regier2016celeste,
@@ -145,7 +149,9 @@ archivePrefix = "arXiv",
   number={1},
   pages={184--214},
   year={2017},
-  publisher={SIAM}
+  publisher={SIAM},
+  url={https://doi.org/10.1137/15M1041250},
+  doi={10.1137/15M1041250}
 }
 
 @article{packwood2016universal,
@@ -156,7 +162,8 @@ archivePrefix = "arXiv",
   number={16},
   pages={164109},
   year={2016},
-  publisher={AIP Publishing}
+  publisher={AIP Publishing},
+  doi = {10.1063/1.4947024}
 }
 
 @article{braun2017effect,
@@ -164,4 +171,11 @@ archivePrefix = "arXiv",
   author={Braun, Julian and Buze, Maciej and Ortner, Christoph},
   journal={arXiv preprint arXiv:1710.07708},
   year={2017}
+}
+
+@misc{johnson2018nlopt,
+author = {Johnson, Steven G.},
+title = {The NLopt nonlinear-optimization package},
+url = {http://ab-initio.mit.edu/nlopt},
+year = {2008--2018}
 }
