### 2018-03-15 739b291
A brief notice for potential contributors, and brief documentation of Particle Swarm. (#548)
