### 2018-03-15 #615 @pkofod
> Review
>
> The Optim package for Julia is well known in the community, and it's now considered an essential part of it. The package follows the good practices expected by JOSS, and I have only a few minor questions, corrections, and suggestions to make:
>
>     Clarify that optimization means mathematical optimization. Possibly in the title.
>     Shouldn’t John Myles White be an author? I’m not familiar with the package history, nor with ownership “rules”, but since he’s made major contributions, it warrants the question. Answering this will probably clarify Tim Holy's situation as well.
>     What’s the difference between “Newton with trust region” and “Hessian-vector with trust region”?
>     ParticleSwarm doesn’t appear to be documented.
>     There are no clear community guidelines, as requested by JOSS. In my opinion, indicating the issues tab for reporting issues, problems and discussing contributions and indicating the gitter or discourse channel for help should be enough.
>     At least one reference needs DOI.
>
> Best,

@abelsiqueira Thanks again for your feedback. We've tried to address your concerns in the two latest PRs: https://github.com/JuliaNLSolvers/Optim.jl/pull/546 https://github.com/JuliaNLSolvers/Optim.jl/pull/548 . We would appreciate if you could verify that the changes cover your concerns. The authorship concerns will depend partially upon what @labarba has to say on the matter based on my earlier comment.
