### 2018-03-15 #615 @labarba
Normally, we expect all persons who've made substantial contributions to the software to be in the list of authors. If both @johnmyleswhite and @timholy have been inactive in the project for a long time, and they approve of the JOSS submission not listing them as authors, then we can accept that. But their names should still be listed in an original copyright notice, as per the license conditions. (I didn't check if this is already the case.)
