### 2018-03-15 #615 @pkofod
> Normally, we expect all persons who've made substantial contributions to the software to be in the list of authors. If both @johnmyleswhite and @timholy have been inactive in the project for a long time, and they approve of the JOSS submission not listing them as authors, then we can accept that. But their names should still be listed in an original copyright notice, as per the license conditions. (I didn't check if this is already the case.)

The LICENSE.md file mentions "John Myles White and other contributors"

> and they approve of the JOSS submission not listing them as authors, then we can accept that

Yes, but let me repeat that this is not due to an explicit wish to exclude them. I think we will try to go the route of waiting for the review of @ahwillia , and then when everything is good to go besides the authorship, we will ship the draft off to facebook's legal department and see what they say. In Tim's case it's not so much that we've been looking for an explicit approval to exclude him, we've rather been unable to receive consent to publish what you see in the draft with his name attached to it. I figure it's a busy time in general, and especially so for Julia core contributors, so maybe we'll hear back before this review is over.
