### 2018-03-15 #615 @ahwillia
Does the year of the LICENSE matter? It is listed as 2012 and John Myles White is the only name listed. I think it would make sense to add a few more names there. Perhaps "John Myles White, JuliaNLSolvers organization, and other contributors" ?

I might be wrong, but I seem to remember the package being nearly entirely re-written by @pkofod at some point. So to me it seems fine that John doesn't necessarily need to be on the author list.

Overall the package is extremely mature and extensively used in the Julia community. So it is a no-brainer that this should be accepted and published. I struggle to come up with suggestions and feedback. Here a couple of small comments:

* There is a checkbox for me at the top of the page which says that all references need a DOI? If this is a requirement, it seems like some of the references in the paper are missing them.

* I think this was mentioned before, but I think it would be nice to have a short section in the documentation and/or the paper to discuss what other libraries are available to users. It seems like you have recently added support for manifold optimization (really cool!!) and to my knowledge this isn't available in `scipy.optimize` (but see [pymanopt](https://pymanopt.github.io/)). There is a nice discussion about the advantages of Julia in the documentation, but it might still be useful to mention these other libraries by name in the paper?

* Similar to above, it might be nice to have a citation to `Convex.jl` and `JuMP.jl` somewhere in the paper. I know `Optim.jl` is quite different than those packages, but it would be nice to mention them and briefly explain to new users how they are different.

* This is really minor and not necessary - but in addition to `grid_search` (a zeroth order multivariate solver) you might consider adding `rand_search`, see [Bergstra & Bengio (2012)](http://www.jmlr.org/papers/volume13/bergstra12a/bergstra12a.pdf). I expect this would be trivial to implement and it could be useful.

* Out of curiosity, is there a nice API to do a nested optimization? Let's say I wanted to use `rand_search` to tune a few hyperparameters of a complex model. For each set of hyperparameters I used a more sophisticated method (e.g. L-BFGS). That would be a super rad use case of this package!

I am happy to sign off on the package as is. Great work putting this together!
