### 2018-03-15 #615 @labarba
We do it like this, reflecting major release years, with author lists changing according to level of contribution:

(c) 2012 Founding Author, Second Founder
(c) 2014 Founding Author, Second Founder, New Contributor
(c) 2018 Main Contributor, New Contributor, Another Contributor, Founding Author, Second Founder

You just keep adding (c) lines below, and over time some authors may drop.

Then the license text comes below.
