### 2018-03-15 #615 @pkofod
Thanks for the feedback @ahwillia

> There is a checkbox for me at the top of the page which says that all references need a DOI? If this is a requirement, it seems like some of the references in the paper are missing them.

I think some of these have already been adressed. Is it possible to generate a new paper pdf @labarba ?

For the rest,

> I think this was mentioned before, but I think it would be nice to have a short section in the documentation and/or the paper to discuss what other libraries are available to users. It seems like you have recently added support for manifold optimization (really cool!!) and to my knowledge this isn't available in scipy.optimize (but see pymanopt). There is a nice discussion about the advantages of Julia in the documentation, but it might still be useful to mention these other libraries by name in the paper?

I am totally fine mentioning scipy.optimize (and JuMP and Convex), but it is not totally clear to me where to draw the line. Should R's optim also be mentioned then?
