### 2018-03-15 #615 @ahwillia
> I am totally fine mentioning scipy.optimize (and JuMP and Convex), but it is not totally clear to me where to draw the line. Should R's optim also be mentioned then?

It wouldn't hurt, I suppose. Is the biggest different between R's optim, `scipy.optim`, and Julia's Optim the language? Or are there other differences that are worth mentioning?

Those are the questions I could see a new user wondering.
