### 2018-03-15 #615 @pkofod
> It wouldn't hurt, I suppose. Is the biggest different between R's optim, scipy.optim, and Julia's Optim the language? Or are there other differences that are worth mentioning?
>
> Those are the questions I could see a new user wondering.

Gotcha. I'll talk to @anriseth and come up with some changes to address your comments (the other ones as well). Thanks.
