### 2018-03-15 JuliaNLSolvers/Optim.jl#550 @anriseth
Feature request: rand_search
This was mentioned by @ahwillia in the JOSS review. Up for grabs for anybody who would like to implement a solver :)

Ref https://github.com/openjournals/joss-reviews/issues/615#issuecomment-373460590

>  in addition to grid_search (a zeroth order multivariate solver) you might consider adding rand_search, see [Bergstra & Bengio (2012)](http://www.jmlr.org/papers/volume13/bergstra12a/bergstra12a.pdf). I expect this would be trivial to implement and it could be useful.
