### 2018-03-15 JuliaNLSolvers/Optim.jl#551 @anriseth
API/Examples of nested optimization?
Mentioned by @ahwillia in the [JOSS review](https://github.com/openjournals/joss-reviews/issues/615#issuecomment-373460590).

> Is there a nice API to do a nested optimization? Let's say I wanted to use rand_search to tune a few hyperparameters of a complex model. For each set of hyperparameters I used a more sophisticated method (e.g. L-BFGS). That would be a super rad use case of this package!

So this seems like a use case that can be relevant for statistical/machine learning.

If anybody wants to have a go at providing an example/documentation/necessary code to do nested optimization, that's much appreciated :)
