### 2018-03-16 a146a52
JOSS. Add python references and update LICENSE (#549)
* Update LICENSE

* Mention scipy.optiize and pymanopt
