+++ joss/paper.bib
@@ -174,8 +174,31 @@ archivePrefix = "arXiv",
 }
 
 @misc{johnson2018nlopt,
-author = {Johnson, Steven G.},
-title = {The NLopt nonlinear-optimization package},
-url = {http://ab-initio.mit.edu/nlopt},
-year = {2008--2018}
+  author = {Johnson, Steven G.},
+  title = {The NLopt nonlinear-optimization package},
+  url = {http://ab-initio.mit.edu/nlopt},
+  year = {2008}
+}
+
+@article{dunning2017jump,
+  title={JuMP: A modeling language for mathematical optimization},
+  author={Dunning, Iain and Huchette, Joey and Lubin, Miles},
+  journal={SIAM Review},
+  volume={59},
+  number={2},
+  pages={295--320},
+  year={2017},
+  publisher={SIAM},
+  doi={10.1137/15M1020575}
+}
+
+@article{townsend2016pymanopt,
+  author  = {Townsend, James and Koep Niklas and Weichwald, Sebastian},
+  title   = {Pymanopt: A {Python} Toolbox for Optimization on Manifolds using Automatic Differentiation},
+  journal = {Journal of Machine Learning Research},
+  year    = {2016},
+  volume  = {17},
+  number  = {137},
+  pages   = {1-5},
+  url     = {http://jmlr.org/papers/v17/16-177.html}
 }
