+++ joss/paper.md
@@ -34,15 +34,26 @@ calculated using automatic differentiation or finite difference
 methods.  The main focus of the package has currently been on
 unconstrained optimization, however, box-constrained optimization is
 supported, and a more comprehensive support for constraints is
-underway.  Similar to Optim, the C library
+underway.
+
+Similar to Optim, the C library
 [NLopt](http://ab-initio.mit.edu/nlopt) [@johnson2018nlopt] contains a
-collection of nonlinear optimization routines. Within the Julia
-community, the packages
+collection of nonlinear optimization routines. In Python,
+[scipy.optimize](https://docs.scipy.org/doc/scipy/reference/optimize.html)
+supports many of the same algorithms as Optim does, and
+[Pymanopt](https://pymanopt.github.io/) [@townsend2016pymanopt] is a
+toolbox for manifold optimization.
+Within the Julia community, the packages
 [BlackBoxOptim.jl](https://github.com/robertfeldt/BlackBoxOptim.jl)
 and
 [Optimize.jl](https://github.com/JuliaSmoothOptimizers/Optimize.jl)
 provide optimization capabilities focusing on derivative-free and
 large-scale smooth problems respectively.
+The packages [Convex.jl](https://github.com/JuliaOpt/Convex.jl) and
+[JuMP.jl](https://github.com/JuliaOpt/JuMP.jl) [@dunning2017jump] define
+modelling languages for which users can formulate optimization problems.
+In contrast to the previously mentioned optimization codes, Convex and JuMP
+work as abstraction layers between the user and solvers from a other packages.
 
 ## Optimization routines
 As of version 0.14, the following optimization routines are available.
