### 2018-03-16 JuliaNLSolvers/Optim.jl#551 @pkofod
The only thing I'm a bit curious about, is how you would pass the first set of parameters in a generic fashion. The simplest way that I can think of to do right now would be something like:

https://github.com/pkofod/Misc/blob/master/Untitled38.ipynb

Basically I use SimulatedAnnealing to find a starting point for BFGS that gives me the lowest objective value, such that if I start BFGS at the minimizer returned by SA, I get exactly 0.0 as the minimum, but if I start BFGS at the point where I start SimulatedAnnealing, it's "only" around 1e-25...

I think that it might actually be easy enough that we don't need a specific interface for it, but we do need a tutorial for it (I just need a better example where we're optimizing over some hyperparameters on the outside).
