### 2018-03-16 JuliaNLSolvers/Optim.jl#550 @pkofod
I'll do it,  according to an interview like
```
optimize(f, lb, ub, RandSearch())
```
does that sound reasonable @ahwillia?
