### 2018-03-16 JuliaNLSolvers/Optim.jl#551 @mohamed82008
This is for unconstrained smooth minimization.
