### 2018-03-16 JuliaNLSolvers/Optim.jl#551 @pkofod
> This is for unconstrained smooth minimization.

What do you mean?
