### 2018-03-16 JuliaNLSolvers/Optim.jl#551 @antoine-levitt
Smooth yes, unconstrained no: even if the minimization is constrained on y, the statement holds (∂g/∂y is non-zero, but dy/dx is in its kernel)
