### 2018-03-16 JuliaNLSolvers/Optim.jl#551 @pkofod
> An interesting trivia about nested optimization that is usually very useful and might be good to point out in a tutorial is that the gradient of the outer objective function is readily available: if f(x) = min_y g(x,y), then f'(x) = ∂g/∂x + ∂g/∂y dy/dx = ∂g/∂x (if the inner optimizer is well-converged so that ∂g/∂y = 0). This is much quicker than using finite differences or autodiff (which would require differentiating the minimization process)

Will remember to point that out.
