### 2018-03-16 JuliaNLSolvers/Optim.jl#551 @mohamed82008
> (∂g/∂y is non-zero, but dy/dx is in its kernel)
May you elaborate? dy/dx is not explicit since it depends on how x affects the optimal solution.
