### 2018-03-16 #615 @anriseth
We have updated the license and addressed @ahwillia's points (scipy.optimize, JuMP, and Convex.jl). Thanks for the feedback!
