### 2018-03-16 JuliaNLSolvers/Optim.jl#551 @antoine-levitt
That's right, but the beauty of it is that it doesn't matter. The statement is that ∂g/∂y is zero on any admissible direction for y (if it wasn't, you could decrease the objective by going in that direction a bit). Since dy/dx is an admissible direction by definition, that term is zero. You can play the same trick with Lagrange multipliers if you prefer that formalism, see for instance the wikipedia page for the Hellmann-Feynman theorem, applied to the case of the minimization of <y, H(x) y> s.t. ||y||=1 where H is a hermitian matrix depending on x. (of course, this is all assuming that the minimization problem is non-degenerate, and therefore by the IFT that y(x) is smooth)
