### 2018-03-16 JuliaNLSolvers/Optim.jl#551 @anriseth
> (∂g/∂y is non-zero, but dy/dx is in its kernel)

I have an example, which does not seem to adhere to this (where am I going wrong?)
`g(x,y) = (x-y)^2` with constraint `y>= 2x`.
Then `y(x) = 2x`, `∂g/∂y = -2(x-y)`, but `dy/dx = 2`, which is not in the kernel of `∂g/∂y |_{y=2x} = 2x`.
