### 2018-03-16 JuliaNLSolvers/Optim.jl#551 @antoine-levitt
Yes sorry, I'm assuming the constraint does not depend on x (if it does I think the calculation can be adapted, you just have to include the Lagrange multiplier of the constraint).
