### 2018-03-16 JuliaNLSolvers/Optim.jl#551 @mohamed82008
If the solution doesn't change with x, that makes sense, but if it does, I doubt this can be directly done. Bender's decomposition for instance handles LPs where linear constraints depend on separable variables, which is a special case of the problem described here.
