### 2018-03-16 JuliaNLSolvers/Optim.jl#551 @antoine-levitt
Unless I'm mistaken that still works. Assume a single constraint h(x,y) = 0, then at the minimum dg/dy + λ dh/dy = 0, so dg/dy dy/dx = - λ dh/dy dy/dx = + λ dh/dx, so you just have to compute the derivative of the constraint with respect to x at the minimum, and the Lagrange multiplier.
