### 2018-03-16 JuliaNLSolvers/Optim.jl#551 @mohamed82008
Hmm, lambda itself could be a function of x!
