### 2018-03-16 JuliaNLSolvers/Optim.jl#551 @mohamed82008
Actually you never assumed lambda was constant, my bad. I think that makes sense.
