### 2018-03-16 JuliaNLSolvers/Optim.jl#551 @antoine-levitt
Of course it does, that's the point of Lagrange multipliers :p that's actually a slightly generalized version of the last equation in https://en.wikipedia.org/wiki/Lagrange_multiplier#Interpretation_of_the_Lagrange_multipliers ("a Lagrange multiplier is the change in the optimal value of the objective function (profit) due to the relaxation of a given constraint")
