### 2018-03-16 JuliaNLSolvers/Optim.jl#550 @ahwillia
Yep!
