### 2018-03-16 JuliaNLSolvers/Optim.jl#551 @ahwillia
Great discussion! I just want to point one thing out briefly:

> the gradient of the outer objective function is readily available

In the applications I'm thinking about this gradient either doesn't exist or is too expensive to compute. For example, if you are optimizing a matrix decomposition problem, or some problem with a discrete number of latent states / clusters. In these cases the hyperparameter is a discrete variable.

Another use case is fitting a deep net via stochastic gradient descent with a fixed learning rate (or even better, a learning rate that decreases according to some schedule). In these cases, it would be very expensive to compute/approximate the necessary gradient.

The second example is maybe outside the scope of what Optim is interested in... But maybe there are others?
