### 2018-03-16 JuliaNLSolvers/Optim.jl#551 @mohamed82008
> Unless I'm mistaken that still works. Assume a single constraint h(x,y) = 0, then at the minimum dg/dy + λ dh/dy = 0, so dg/dy dy/dx = - λ dh/dy dy/dx = + λ dh/dx

On a second less sleepy thought, the above derivation is slightly off. In the lower level problem, minimizing g(x,y) with respect to y given x is the goal. These conditions will be satisfied at the optimal y* if the point y is a KKT point.
-∂g/∂y + λ ∂h/∂y = 0

The goal of the higher level optimization problem is to minimize g(x,y*(x)) where y*(x) is the implicit function that returns the optimal solution y* to the lower level problem given x. The full derivative of g with respect to x is ∂g/∂x + ∂g/∂y* dy*/dx. Since y*(x) is not explicit, one would ideally want to eliminate or find dy*/dx without going through y*(x). Since at the optimal solution of the lower level problem y = y*(x) and g(x,y) = g(x,y*(x)), let's overload y now and treat it as a function in the higher level problem. This relationship still holds -∂g/∂y + λ ∂h/∂y = 0.

Multiplying by dy/dx, -∂g/∂y dy/dx + λ ∂h/∂y dy/dx = 0. Adding suitably weighted components of ∂g/∂y and ∂h/∂x on both sides, -(∂g/∂x + ∂g/∂y dy/dx) + λ (∂h/∂x + ∂h/∂y dy/dx) = -∂g/∂x + λ ∂h/∂x. This reduces to -dg/dx + λ dh/dx = -∂g/∂x + λ ∂h/∂x. So we have a relationship between dg/dx and dh/dx both of which are unknown. This means we need another equation to help us find both since we have no way of computing the full derivative of the constraint or the objective unless it is a function of x only which wouldn't be a nested problem anymore, or it doesn't depend on x at all, which is the case where the solution space does not depend on x.

Did I go wrong somewhere?
