### 2018-03-16 JuliaNLSolvers/Optim.jl#551 @mohamed82008
> The second example is maybe outside the scope of what Optim is interested in... But maybe there are others?
I don't see what stops you from defining the parameters to the lower level optimizer as variables in the higher level function. The only thing is that you have to stick to zero order optimization methods for the higher level problem, use auto-diff through the optimization algorithm, or find an explicit form of the derivative with respect to the hyper-parameter as in this paper: https://arxiv.org/abs/1703.04782.
