### 2018-03-17 JuliaNLSolvers/Optim.jl#551 @mohamed82008
The only case I think Optim will have a hard time handling is the discrete parameter case, since Optim only does continuous optimization.
