### 2018-03-17 JuliaNLSolvers/Optim.jl#551 @pkofod
In the case that was the motivation for this issue, we could support diskrete values as long as there’s a set of diskrete numbers we can sample from. Can you reference some simple tuning problems se could try @ahwillia?
