### 2018-03-17 JuliaNLSolvers/Optim.jl#551 @antoine-levitt
@mohamed82008 I was too quick, and partial derivatives are awkward to type, but you have to use d/dx h(x, y(x)) = 0 (because h(x,y(x)) = 0 for all x), and so ∂h/∂x + ∂h/∂y dy/dx = 0. The end formula only has ∂h/∂x in it, which is easy to compute. In the special case where the constraint is of the form h(y)=x, you get the interpretation of Lagrange multipliers as marginal costs.

@ahwillia Discrete optimization and stochastic optimization are indeed tricky; the general idea which I'm yet to see a counter-example of is that whenever a gradient really makes sense it can be computed efficiently, even for complicated functions with other loops inside.
