### 2018-03-17 JuliaNLSolvers/Optim.jl#551 @mohamed82008
@antoine-levitt Interesting! That's true for equality constraints though, for active inequality constraints of the form h(x,y) <= 0, the best guarantee we have at the top level problem is that h(x,y(x)) can only decrease by changing x, so dh/dx <= 0, unless one can prove that the constraint will remain active by changing x, in which case it is equivalent to the equality version.

However, by complimentarity slackness, λ(x) h(x,y(x)) = 0 -> dλ/dx h(x,y(x)) + λ(x) dh/dx = 0 -> λ(x) dh/dx = -dλ/dx h(x,y(x)). Multiplying by λ(x) -> (λ(x))^2 dh/dx = -dλ/dx λ(x) h(x,y(x)) = 0. Therefore for any x, either λ(x) = 0 or dh/dx = 0. So one can still eliminate the second term in -dg/dx + λ dh/dx = -∂g/∂x + λ ∂h/∂x -> -dg/dx = -∂g/∂x + λ ∂h/∂x.

The only missing but not so important piece I think is when λ = 0, I am not sure how to find dh/dx. I think the best guarantee we have is that dh/dx <= 0. Not that it matters much anyways 😆. This discussion is very useful for my work so thanks a lot!!!
