### 2018-03-17 JuliaNLSolvers/Optim.jl#551 @antoine-levitt
I'm very rusty on inequality constrained optimization, but the only problematic part as you say is when an active constraint becomes inactive when x moves a bit. In this case it's not obvious that everything stays smooth and the gradient is well-defined (probably this can be ensured with the appropriate qualification conditions, or maybe the function is not differentiable and one has to look at subgradients?), but if it does the result (gradient = ∂g/∂x) is valid by continuity.
