### 2018-03-17 JuliaNLSolvers/Optim.jl#551 @ahwillia
> the general idea which I'm yet to see a counter-example of is that whenever a gradient really makes sense it can be computed efficiently

Often in hyperparameter optimization I don't think it makes sense to use gradients, even if you can compute them. Since you are operating in a relatively low-dimensional space, possibly with a very nonconvex function, it makes sense to do a more exploratory/global optimization than a local gradient-based approach.

> Can you reference some simple tuning problems se could try @ahwillia?

A very, very simple one would be LASSO regression where one does a 1D grid search over the strength of the penalty parameter. Since the problem is convex you can use warm starts for each inner loop. If you want to extend this to a 2D example you could try elastic net regression.

To add a discrete variable, you could try something like sparse PCA (which is basically LASSO penalty on a matrix factorization). In that case you would be tuning a penalty parameter as well as the discrete number of hidden states/components. I don't think the Eckart-Young theorem applies to Sparse PCA, so you have to do hyperparameter search over the number of components/clusters in the model
