### 2018-03-18 #615 @abelsiqueira
Sorry about the absence.

I reviewed the corrections and the concerns are all addressed, though there are just two new things I noticed:

- I remembered [NLP.py](http://pythonoptimizers.github.io/NLP.py/latest/) as another similar high-level language software, it might fit in the comparison as well.
- The authors of the references are jumbled (maybe some style I'm not aware?):
> Second, First, First Second, First Second
