### 2018-03-18 #615 @arfon
> Second, First, First Second, First Second

Yeah, I'm not sure about this. We're just using the default Pandoc/citeproc styles but I agree it's a little weird...
