### 2018-03-19 #615 @labarba
I see that both reviewers have checked off all items in the review checklist.

Can I have confirmation that the reviewers recommend acceptance?
@abelsiqueira @ahwillia ?

Also, can I have confirmation from the authors that they've finished working on it?
@anriseth
