### 2018-03-19 #615 @ahwillia
Yes I recommend acceptance.

On Mar 19, 2018 12:39 PM, "Lorena A. Barba" <notifications@github.com>
wrote:

I see that both reviewers have checked off all items in the review
checklist.

Can I have confirmation that the reviewers recommend acceptance?
@abelsiqueira <https://github.com/abelsiqueira> @ahwillia
<https://github.com/ahwillia> ?

Also, can I have confirmation from the authors that they've finished
working on it?
@anriseth <https://github.com/anriseth>
