### 2018-03-19 #615 @pkofod
> Also, can I have confirmation from the authors that they've finished working on it?
> @anriseth

I know Asbjørn submitted it, but please allow me to reach out to the other potential authors one last time, so get their final comments on wether they want to be credited as co-authors of if the acknowledgement is sufficient.
