### 2018-03-20 #615 @labarba
OK, everyone. Thank you sincerely to @ahwillia and @abelsiqueira for their review.

@arfon -- We're waiting for the authors to do a final effort of adjusting their author list with contributors currently not named, due to their employment restrictions. After they give us the flag, the paper is ready for publication.
