### 2018-03-22 d77f269
Add simulated annealing with bounds. (#553)
* Add simulated annealing with bounds.

* Changes.

* Error on out of bounds.

* Fix tests.
