### 2018-03-22 02e5ce2
Update samin.md
The example was "quenching" because the cooling rate was too fast. This small change obtains a much more accurate solution.
