### 2018-03-22 cfdc407
Continue to track LineSearches master.
