### 2018-03-22 8469ff9
Fix test in samin. (#557)
