+++ test/multivariate/solvers/constrained/samin.jl
@@ -4,5 +4,6 @@
     xtrue = prob.solutions
     f = OptimTestProblems.MultivariateProblems.objective(prob)
     x0 = prob.initial_x
-    optimize(f, x0, x0.-100., x0.+100.0, Optim.SAMIN())
+    res = optimize(f, x0, x0.-100., x0.+100.0, Optim.SAMIN(), Optim.Options(iterations=4000))
+    @test Optim.minimum(res) < 1e-6
 end
