### 2018-03-23 15cb1fc
Addressed merge conflict.
