### 2018-03-23 f22b197
copy -> copyto in newton_trust_region
