### 2018-03-23 72f2759
more detailed samin documentation
