+++ docs/src/algo/samin.md
@@ -3,35 +3,55 @@
 ```julia
 SAMIN(; nt::Int = 5     # reduce temperature every nt*ns*dim(x_init) evaluations
         ns::Int = 5     # adjust bounds every ns*dim(x_init) evaluations
-        rt::T = 0.5     # geometric temperature reduction factor: when temp changes, new temp is t=rt*t
+        rt::T = 0.9     # geometric temperature reduction factor: when temp changes, new temp is t=rt*t
         neps::Int = 5   # number of previous best values the final result is compared to
-        f_tol::T = 1e-8 # the required tolerance level for function value comparisons
-        x_tol::T = 1e-5 # the required tolerance level for x
-        coverage_ok::Bool = false, # increase temperature until parameter space is covered
+        f_tol::T = 1e-12 # the required tolerance level for function value comparisons
+        x_tol::T = 1e-6 # the required tolerance level for x
+        coverage_ok::Bool = false, # if false, increase temperature until initial parameter space is covered
         verbosity::Int = 0) # scalar: 0, 1, 2 or 3 (default = 0).
 ```
 ## Description
 The `SAMIN` method implements the Simulated Annealing algorithm for problems with
-bounds constrains a described in Goffe et. al. (1994) and Goffe (1996). The
-algorithm requires bounds, although these bounds are often set rather wide, and
-are not necessarily meant to reflect constraints in the model, but rather bounds
-that are there to improve the search. If the final `x`s are very close to the boundary,
-it is a good idea to restart the optimizer with wider bounds, unless the bounds
-actually reflect hard constraints on `x`.
+bounds constraints as described in Goffe et. al. (1994) and Goffe (1996). A key control
+parameter is rt, the geometric temperature reduction rate, which should be between zero
+and one. Setting rt lower will cause the algorithm to contract the search space more quickly,
+reducing the run time. Setting rt too low will cause the algorithm to narrow the search
+too quickly, and the true minimizer may be skipped over. If possible, run the algorithm
+multiple times to verify that the same solution is found each time. If this is not the case,
+increase rt. When in doubt, start with a conservative rt, for example, rt=0.95, and allow for
+a generous iteration limit. The algorithm requires lower and upper bounds on the parameters,
+although these bounds are often set rather wide, and are not necessarily meant to reflect
+constraints in the model,but rather bounds that enclose the parameter space. If the final
+`x`s are very close to the boundary (which can be checked by setting verbosity=1), it is a
+good idea to restart the optimizer with wider bounds, unless the bounds actually reflect
+hard constraints on `x`.
 
 ## Example
+This is example shows a successful minimization:
 ```julia
 julia> using Optim, OptimTestProblems
 
 julia> prob = OptimTestProblems.UnconstrainedProblems.examples["Rosenbrock"];
 
-julia> res = Optim.optimize(prob.f, prob.initial_x, fill(-100.0, 2), fill(100.0, 2), SAMIN(rt=0.9), Optim.Options(iterations=20000))
-Results of Optimization AlgorithmResults of Optimization Algorithm
+julia> res = Optim.optimize(prob.f, prob.initial_x, fill(-100.0, 2), fill(100.0, 2), SAMIN(), Optim.Options(iterations=10^6))
+================================================================================
+SAMIN results
+==> Normal convergence <==
+total number of objective function evaluations: 23701
+
+     Obj. value:      0.0000000000
+
+       parameter      search width
[…]
