### 2018-03-23 c147bc2
change some defaults, modify verbosity
* the default rt has been set to 0.9, which is more conservative, and more likely to obtain a correct solution
* the final results are displayed, until the convergence of samin can be integrated into the general scheme. Otherwise, it appears that the algorithm fails to converge, when in fact it does.
