+++ src/multivariate/solvers/constrained/fminbox.jl
@@ -90,13 +90,14 @@ end
 # Default preconditioner for box-constrained optimization
 # This creates the inverse Hessian of the barrier penalty
 function precondprepbox!(P, x, l, u, mu)
-    @. P.diag = 1/(mu*(1/(x-l)^2 + 1/(u-x)^2) + 1)
+    T = eltype(x)
+    @. P.diag = 1/(mu*(1/max(x-l, sqrt(realmin(T)))^2 + 1/max(u-x, sqrt(realmin(T)))^2) + 1)
 end
 
 struct Fminbox{T<:AbstractOptimizer} <: AbstractOptimizer end
 Fminbox() = Fminbox{ConjugateGradient}() # default optimizer
 
-Base.summary(::Fminbox{O}) where {O} = "Fminbox with $(summary(O()))"
+Base.summary(::Fminbox{O}) where {O} = "Fminbox with $(summary(O(T)))"
 
 function optimize(obj,
                   initial_x::AbstractArray{T},
@@ -162,11 +163,11 @@ function optimize(
         thisu = u[i]
 
         if thisx == thisl
-            thisx = 0.99*thisl+0.01*thisu
+            thisx = T(99)/100*thisl+T(1)/100*thisu
             x[i] = thisx
             push!(boundaryidx,i)
         elseif thisx == thisu
-            thisx = 0.01*thisl+0.99*thisu
+            thisx = T(1)/100*thisl+T(99)/100*thisu
             x[i] = thisx
             push!(boundaryidx,i)
         elseif thisx < thisl || thisx > thisu
@@ -213,47 +214,47 @@ function optimize(
         #       in the optimization algorithms will imply a lot of extra work here
         if O == ConjugateGradient || O == Newton
             if linesearch == nothing
-                linesearch = LineSearches.HagerZhang()
+                linesearch = LineSearches.HagerZhang{T}()
             end
             if alphaguess == nothing
-                alphaguess = LineSearches.InitialHagerZhang()
+                alphaguess = LineSearches.InitialHagerZhang{T}()
             end
             _optimizer = ConjugateGradient(eta = eta, alphaguess = alphaguess,
                                            linesearch = linesearch, P = P, precondprep = pcp)
         elseif O == LBFGS
             if linesearch == nothing
-                linesearch = LineSearches.HagerZhang()
+                linesearch = LineSearches.HagerZhang{T}()
             end
             if alphaguess == nothing
-                alphaguess = LineSearches.InitialStatic()
+                alphaguess = LineSearches.InitialStatic{T}()
             end
-            _optimizer = O(alphaguess = alphaguess, linesearch = linesearch, P = P, precondprep = pcp)
+            _optimizer = O(T, alphaguess = alphaguess, linesearch = linesearch, P = P, precondprep = pcp)
         elseif O == BFGS
             if linesearch == nothing
-                linesearch = LineSearches.HagerZhang()
+                linesearch = LineSearches.HagerZhang{T}()
             end
             if alphaguess == nothing
-                alphaguess = LineSearches.InitialStatic()
+                alphaguess = LineSearches.InitialStatic{T}()
             end
-            _optimizer = O(alphaguess = alphaguess, linesearch = linesearch)
+            _optimizer = O(T, alphaguess = alphaguess, linesearch = linesearch)
         elseif O == GradientDescent
             if linesearch == nothing
-                linesearch = LineSearches.HagerZhang()
+                linesearch = LineSearches.HagerZhang{T}()
             end
             if alphaguess == nothing
-                alphaguess = LineSearches.InitialPrevious()
+                alphaguess = LineSearches.InitialPrevious(alpha=one(T), alphamin=zero(T), alphamax=T(Inf))
             end
[…]
