+++ src/multivariate/solvers/zeroth_order/simulated_annealing.jl
@@ -1,11 +1,12 @@
 log_temperature(t::Real) = 1 / log(t)
 
-constant_temperature(t::Real) = 1.0
+constant_temperature(t::T) where T = T(1)
 
 function default_neighbor!(x::AbstractArray, x_proposal::AbstractArray)
     @assert size(x) == size(x_proposal)
+    T = eltype(x)
     for i in 1:length(x)
-        @inbounds x_proposal[i] = x[i] + randn()
+        @inbounds x_proposal[i] = x[i] + randn(T)
     end
     return
 end
@@ -44,8 +45,7 @@ function initial_state(method::SimulatedAnnealing, options, d, initial_x::Abstra
     SimulatedAnnealingState(copy(best_x), 1, best_x, similar(initial_x), value(d), value(d))
 end
 
-function update_state!(nd, state::SimulatedAnnealingState{T}, method::SimulatedAnnealing) where T
-
+function update_state!(nd, state::SimulatedAnnealingState{Tx,T}, method::SimulatedAnnealing) where {Tx,T}
     # Determine the temperature for current iteration
     t = method.temperature(state.iteration)
 
@@ -68,7 +68,7 @@ function update_state!(nd, state::SimulatedAnnealingState{T}, method::SimulatedA
     else
         # If proposal is inferior, we move to it with probability p
         p = exp(-(state.f_proposal - state.f_x_current) / t)
-        if rand() <= p
+        if rand(T) <= p
             copy!(state.x_current, state.x_proposal)
             state.f_x_current = state.f_proposal
         end
