+++ src/types.jl
@@ -37,15 +37,14 @@ function Options(;
         show_trace::Bool = false,
         extended_trace::Bool = false,
         show_every::Integer = 1,
-        callback = nothing,
-        time_limit = NaN)
+        callback::TCallback = nothing,
+        time_limit = NaN) where {TCallback}
     show_every = show_every > 0 ? show_every : 1
     #if extended_trace && callback == nothing
     #    show_trace = true
     #end
-    Options(promote(x_tol, f_tol, g_tol)..., f_calls_limit, g_calls_limit, h_calls_limit,
-        allow_f_increases, Int(iterations), store_trace, show_trace, extended_trace,
-        Int(show_every), callback, time_limit)
+    T = promote_type(typeof(x_tol), typeof(f_tol), typeof(g_tol))
+    return Options{T,TCallback}(promote(x_tol, f_tol, g_tol)..., f_calls_limit, g_calls_limit, h_calls_limit, allow_f_increases, successive_f_tol, Int(iterations), store_trace, show_trace, extended_trace, Int(show_every), callback, time_limit)
 end
 
 function print_header(options::Options)

[…]
