### 2018-03-24 5662b7e
Fix errors from arbitrary precision commit
