### 2018-03-24 ac7f2ca
Working on type stability.
