+++ src/multivariate/solvers/first_order/ngmres.jl
@@ -71,7 +71,7 @@ function NGMRES(;manifold::Manifold = Flat(),
                     alphaguess = LineSearches.InitialPrevious(),
                     linesearch = LineSearches.Static(alpha=1e-4,scaled=true), # Step length arbitrary
                     manifold = manifold),
-                nlpreconopts = Options(iterations = 1, allow_f_increases = true),
+                nlpreconopts = InternalUseOptions(1e-32,1e-32,1e-8,0,0,0,true,0,1),#Options(iterations = 1, allow_f_increases = true),
                 ϵ0 = 1e-12, # ϵ0 = 1e-12  -- number was an arbitrary choice#
                 wmax::Int = 10) # wmax = 10  -- number was an arbitrary choice to match L-BFGS field `m`
     @assert manifold == nlprecon.manifold
@@ -112,7 +112,7 @@ function OACCEL(;manifold::Manifold = Flat(),
                     alphaguess = LineSearches.InitialPrevious(),
                     linesearch = LineSearches.Static(alpha=1e-4,scaled=true), # Step length arbitrary
                     manifold = manifold),
-                nlpreconopts = Options(iterations = 1, allow_f_increases = true),
+                nlpreconopts = InternalUseOptions(1e-32,1e-32,1e-8,0,0,0,true,0,1),# Options(iterations = 1, allow_f_increases = true),
                 ϵ0 = 1e-12, # ϵ0 = 1e-12  -- number was an arbitrary choice
                 wmax::Int = 10) # wmax = 10  -- number was an arbitrary choice to match L-BFGS field `m`
     @assert manifold == nlprecon.manifold
@@ -136,7 +136,7 @@ mutable struct NGMRESState{P,Tx,Te,T} <: AbstractOptimizerState where P <: Abstr
     ξ::Te                     # Storage to create linear system
     curw::Int                 # Counter for current window size
     A::Array{T,2}             # Container for Aα = b
-    b::Vector{T}                     # Container for Aα = b
+    b::Vector{T}              # Container for Aα = b
     xA::Vector{T}             # Container for accelerated step
     k::Int                    # Used for indexing where to put values in the Storage containers
     restart::Bool             # Restart flag
@@ -434,4 +434,7 @@ end
 
 function default_options(method::AbstractNGMRES)
     Dict(:allow_f_increases => true)
+end
+function InternalUseOptions(::AbstractNGMRES) # sets: allow_f_increases::Bool = true
+    InternalUseOptions(1e-32, 1e-32, 1e-8, 0, 0, 0, true)
 end
