### 2018-03-26 24ab253
Handle deprecation warnings for Newton's Trust Region.
