+++ src/multivariate/solvers/second_order/newton_trust_region.jl
@@ -40,7 +40,7 @@ end
 function p_sq_norm(lambda::T, min_i, n, qg, H_eig) where T
     p_sum = zero(T)
     for i = min_i:n
-        p_sum += qg[i]^2 / (lambda + H_eig[:values][i])^2
+        p_sum += qg[i]^2 / (lambda + H_eig.values[i])^2
     end
     p_sum
 end
@@ -81,13 +81,13 @@ function solve_tr_subproblem!(gr,
     # Note that currently the eigenvalues are only sorted if H is perfectly
     # symmetric.  (Julia issue #17093)
     H_eig = eigfact(Symmetric(H))
-    min_H_ev, max_H_ev = H_eig[:values][1], H_eig[:values][n]
+    min_H_ev, max_H_ev = H_eig.values[1], H_eig.values[n]
     H_ridged = copy(H)
 
     # Cache the inner products between the eigenvectors and the gradient.
     qg = similar(gr)
     for i=1:n
-        qg[i] = vecdot(H_eig[:vectors][:, i], gr)
+        qg[i] = vecdot(H_eig.vectors[:, i], gr)
     end
 
     # These values describe the outcome of the subproblem.  They will be
@@ -100,7 +100,7 @@ function solve_tr_subproblem!(gr,
         # No shrinkage is necessary: -(H \ gr) is the minimizer
         interior = true
         reached_solution = true
-        s[:] = -(H_eig[:vectors] ./ H_eig[:values]') * H_eig[:vectors]' * gr
+        s[:] = -(H_eig.vectors ./ H_eig.values') * H_eig.vectors' * gr
         lambda = zero(T)
     else
         interior = false
@@ -108,7 +108,7 @@ function solve_tr_subproblem!(gr,
         # The hard case is when the gradient is orthogonal to all
         # eigenvectors associated with the lowest eigenvalue.
         hard_case_candidate, min_H_ev_multiplicity =
-            check_hard_case_candidate(H_eig[:values], qg)
+            check_hard_case_candidate(H_eig.values, qg)
 
         # Solutions smaller than this lower bound on lambda are not allowed:
         # they don't ridge H enough to make H_ridge PSD.
@@ -137,10 +137,10 @@ function solve_tr_subproblem!(gr,
                 # I don't think it matters which eigenvector we pick so take
                 # the first.
                 for i=1:n
-                    s[i] = tau * H_eig[:vectors][i, 1]
+                    s[i] = tau * H_eig.vectors[i, 1]
                     for k=(min_H_ev_multiplicity + 1):n
                         s[i] = s[i] +
-                               qg[k] * H_eig[:vectors][i, k] / (H_eig[:values][k] + lambda)
+                               qg[k] * H_eig.vectors[i, k] / (H_eig.values[k] + lambda)
                     end
                 end
             end
