### 2018-03-27 921fd8a
0.7 test fixes; does not pass yet.
