+++ src/multivariate/precon.jl
@@ -34,7 +34,7 @@ precondprep!(P, x) = nothing
 #
 
 # out =  P^{-1} * A
-A_ldiv_B!(out, ::Nothing, A) = copy!(out, A)
+A_ldiv_B!(out, ::Nothing, A) = copyto!(out, A)
 # A' * P B
 dot(A, ::Nothing, B) = vecdot(A, B)
 
@@ -45,7 +45,7 @@ dot(A, ::Nothing, B) = vecdot(A, B)
 #      unfortunately, Base does not implement
 #      A_ldiv_B!(a, P, b) or A_mul_B! for this type, so we do it by hand
 #      TODO: maybe implement this in Base
-A_ldiv_B!(out::Array, P::Diagonal, A::Array) = copy!(out, A ./ P.diag)
+A_ldiv_B!(out::Array, P::Diagonal, A::Array) = copyto!(out, A ./ P.diag)
 dot(A::Array, P::Diagonal, B::Array) = vecdot(A, P.diag .* B)
 
 #####################################################
@@ -55,7 +55,7 @@ dot(A::Array, P::Diagonal, B::Array) = vecdot(A, P.diag .* B)
 mutable struct InverseDiagonal
    diag
 end
-A_ldiv_B!(out::Array, P::InverseDiagonal, A::Array) = copy!(out, A .* P.diag)
+A_ldiv_B!(out::Array, P::InverseDiagonal, A::Array) = copyto!(out, A .* P.diag)
 dot(A::Array, P::InverseDiagonal, B::Vector) = vecdot(A, B ./ P.diag)
 
 #####################################################
@@ -64,4 +64,4 @@ dot(A::Array, P::InverseDiagonal, B::Vector) = vecdot(A, B ./ P.diag)
 #     > A_ldiv_B! is about to be moved to Base, so we need a temporary hack
 #     > A_mul_B! is already in Base, which defines `dot`
 #  nothing to do!
-A_ldiv_B!(x, P::AbstractMatrix, b) = copy!(x, P \ b)
+A_ldiv_B!(x, P::AbstractMatrix, b) = copyto!(x, P \ b)
