+++ src/multivariate/solvers/first_order/cg.jl
@@ -150,7 +150,7 @@ function update_state!(d, state::ConjugateGradientState, method::ConjugateGradie
         # Search direction is predetermined
 
         # Maintain a record of the previous gradient
-        copy!(state.g_previous, gradient(d))
+        copyto!(state.g_previous, gradient(d))
 
         # Determine the distance of movement along the search line
         lssuccess = perform_linesearch!(state, method, ManifoldObjective(method.manifold, d))
@@ -182,7 +182,7 @@ function update_state!(d, state::ConjugateGradientState, method::ConjugateGradie
         etak::eltype(state.x) = method.eta * vecdot(state.s, state.g_previous) / dPd
         state.y .= gradient(d) .- state.g_previous
         ydots = vecdot(state.y, state.s)
-        copy!(state.py, state.pg)        # below, store pg - pg_previous in py
+        copyto!(state.py, state.pg)        # below, store pg - pg_previous in py
         A_ldiv_B!(real_to_complex(d,state.pg), method.P, real_to_complex(d,gradient(d)))
         state.py .= state.pg .- state.py
         betak = (vecdot(state.y, state.pg) - vecdot(state.y, state.py) * vecdot(gradient(d), state.s) / ydots) / ydots
