+++ test/general/counter.jl
@@ -76,7 +76,8 @@
         n = length(x)
         H = Matrix{Float64}(undef, n, n)
         h!(H, x)
-        out .= H * v
+        (out .= H * v; nothing)
+        out
     end
     begin
         solver = Optim.KrylovTrustRegion()
