+++ src/multivariate/solvers/first_order/ngmres.jl
@@ -215,17 +215,17 @@ function initial_state(method::AbstractNGMRES, options, d, initial_x::AbstractAr
 
     n = length(nlpreconstate.x)
     wmax = method.wmax
-    X = Array{T}(n, wmax)
-    R = Array{T}(n, wmax)
-    Q = Array{T}(wmax, wmax)
+    X = Array{T}(undef, n, wmax)
+    R = Array{T}(undef, n, wmax)
+    Q = Array{T}(undef, wmax, wmax)
     ξ = if typeof(method) <: OACCEL
-        Array{T}(wmax, 2)
+        Array{T}(undef, wmax, 2)
     else
-        Array{T}(wmax)
+        Array{T}(undef, wmax)
     end
 
-    copy!(view(X,:,1), nlpreconstate.x)
-    copy!(view(R,:,1), gradient(d))
+    copyto!(view(X,:,1), nlpreconstate.x)
+    copyto!(view(R,:,1), gradient(d))
 
     _updateQ!(Q, 1, 1, X, R, method)
 
@@ -243,13 +243,13 @@ function initial_state(method::AbstractNGMRES, options, d, initial_x::AbstractAr
                 Q,
                 ξ,
                 1,                        # curw
-                Array{T}(wmax, wmax),     # A
-                Array{T}(wmax),           # b
+                Array{T}(undef, wmax, wmax),     # A
+                Array{T}(undef, wmax),           # b
                 vec(similar(initial_x)),  # xA
                 0,                        # iteration counter
                 false,                    # Restart flag
                 options.g_tol,            # Exit tolerance check after nonlinear preconditioner apply
-                Array{T}(wmax),           # subspacealpha
+                Array{T}(undef, wmax),           # subspacealpha
                 @initial_linesearch()...)
 end
 
@@ -266,7 +266,7 @@ end
 
 function update_state!(d, state::NGMRESState{X,T}, method::AbstractNGMRES) where X where T
     # Maintain a record of previous position, for convergence assessment
-    copy!(state.x_previous_0, state.x)
+    copyto!(state.x_previous_0, state.x)
     state.f_x_previous_0 = value(d)
 
     state.k += 1
@@ -367,7 +367,7 @@ function update_state!(d, state::NGMRESState{X,T}, method::AbstractNGMRES) where
     Update x_previous and f_x_previous to be the values at the beginning
     of the N-GMRES iteration. For convergence assessment purposes.
     =#
-    copy!(state.x_previous, state.x_previous_0)
+    copyto!(state.x_previous, state.x_previous_0)
     state.f_x_previous = state.f_x_previous_0
 
     lssuccess == false # Break on linesearch error
@@ -387,8 +387,8 @@ function update_g!(d, state, method::AbstractNGMRES)
     end
     j = mod(state.k, method.wmax) + 1
 
-    copy!(view(state.X,:,j), vec(state.x))
-    copy!(view(state.R,:,j), vec(gradient(d)))
+    copyto!(view(state.X,:,j), vec(state.x))
+    copyto!(view(state.R,:,j), vec(gradient(d)))
 
     for i = 1:state.curw
         _updateQ!(state.Q, i, j, state.X, state.R, method)
