+++ test/multivariate/solvers/second_order/krylov_trust_region.jl
@@ -49,9 +49,10 @@ end
       if prob.istwicedifferentiable
             hv!(storage::Vector, x::Vector, v::Vector) = begin
                 n = length(x)
-                H = Matrix{Float64}(n, n)
+                H = Matrix{Float64}(undef, n, n)
                 MVP.hessian(prob)(H, x)
-                storage .= H * v
+                (storage .= H * v; nothing)
+                storage
             end
             fg!(g::Vector, x::Vector) = begin
                 MVP.gradient(prob)(g,x)
