+++ test/runtests.jl
@@ -4,7 +4,10 @@ using OptimTestProblems.MultivariateProblems
 using Compat.Test
 using Suppressor
 
-using Compat.LinearAlgebra, Compat.Random
+using Compat.LinearAlgebra, Compat.SparseArrays, Compat.Random
+using Distributed: clear!
+
+import NLSolversBase
 
 import LineSearches, ForwardDiff
 
@@ -97,7 +100,7 @@ function run_optim_tests(method; convergence_exceptions = (),
         end
         show_name && print_with_color(:green, "Problem: ", name, "\n")
         # Look for name in the first elements of the iteration_exceptions tuples
-        iter_id = find(n[1] == name for n in collect(iteration_exceptions))
+        iter_id = findall(n -> n[1] == name, iteration_exceptions)
         # If name wasn't found, use default 1000 iterations, else use provided number
         iters = length(iter_id) == 0 ? 1000 : iteration_exceptions[iter_id[1]][2]
         # Construct options
