### 2018-03-28 #615 @pkofod
I think the constructive way forward is to publish as is with the acknowledgement of John Myles White and Tim Holy's contributions early in Optim.jl's life. As mentioned above, this was suggested by JWM in private correspondence, but we have still not heard back from Tim since our very early communication about making a submission to JOSS in January through e-mail, github pings, or otherwise. @anriseth @labarba @arfon
