### 2018-03-28 acee950
Added option to specialize on functions.
