### 2018-03-30 e47e3b9
A few test fixes.
