+++ test/multivariate/precon.jl
@@ -8,7 +8,7 @@
 @testset "Preconditioning" begin
     plap(U; n=length(U)) = (n-1) * sum((0.1 .+ diff(U).^2).^2) - sum(U) / (n-1)
     plap1(U; n=length(U), dU = diff(U), dW = 4 .* (0.1 .+ dU.^2) .* dU) =
-                            (n - 1) .* ([0.0; dW] .- [dW; 0.0]) .- one(U) / (n-1)
+                            (n - 1) .* ([0.0; dW] .- [dW; 0.0]) .- fill!(similar(U),1) / (n-1)
     precond(x::Vector) = precond(length(x))
     precond(n::Number) = spdiagm( -1 => -ones(n-1), 0 => 2*ones(n), 1 => -ones(n-1) ) * (n+1)
     f(X) = plap([0;X;0])
