### 2018-04-01 #615 @labarba
Looks good — small typo in the Acknowledgements:

> We would also like to thank everyone who _has_ contributed…
