### 2018-04-02 #615 @pkofod
https://github.com/JuliaNLSolvers/Optim.jl/pull/565 the typo is fixed on mastee
