### 2018-04-02 #615 @labarba
@arfon : This submission is ready for acceptance.

@anriseth : We now need you to make a new release, upload the revised software to your DOI-granting data/software repository, and post the DOI here.
