### 2018-04-02 #615 @pkofod
Alright, but then the v0.14 self-reference in the paper needs to be changed as well, right?
