+++ versions.js
@@ -16,6 +16,7 @@ var DOC_VERSIONS = [
   "v0.7.1",
   "v0.7.0",
   "v0.6.0",
+  "v0.14.1",
   "v0.14.0",
   "v0.13.0",
   "v0.12.0",
