### 2018-04-04 a921806
Make chunk number accessible for forward diff once differentiable objects.
