### 2018-04-04 e0d2980
Turn tests back on.
