### 2018-04-04 a95f3f1
Track LineSearches master. (#543)
* Track LineSearches master.

* Continue to track LineSearches master.

* Move alpha.
